fix: architecture review — 8 findings (C1, H1-H6, M3, M4) + execution engine hardening
Architecture review fixes (ponytail — minimal change, maximum impact):

Phase 0 — Zero-risk type/annotation fixes:
- M4: Fix _OrderFilled → OrderFilled type annotation in backtest.py
- H1: Fix max_order_value float → Decimal in capabilities.py
- M3: Remove dead store in plot() in market.py

Phase 1 — Pure additions (low risk):
- C1: Bound _applied_fills set at 50K to prevent memory leak (engine.py)
- H5: Add Upstox fill-price override via _stream_order_from_row (mirrors Dhan)
- H6: Add SEBI (₹20/crore) + stamp duty (0.003%) to canonical FeeBreakdown
- H2: Deduplicate _bucketize to delegate to _aggregate in market.py

Phase 2 — Low-medium risk:
- H4: Change OrderResult.status from str → OrderStatus with safe coercion
  downstream: extension.py, client_shared.py, test mocks updated

Deferred (zero runtime impact or breaking-change risk):
- C2 (bus data race), H3 (PlaceOrderCommand CQRS), H7 (instrument_id default),
  M1 (disclosed_quantity), M2 (InstrumentType deprecated), M3 full (plot SRP)

Also includes: execution engine hardening, backtest engine improvements,
analytics/indicators enhancements, strategy engine extensions, walk-forward
optimization, datalake/corporate actions, session lifecycle improvements,
broker WS dedup, fee model expansion, and comprehensive test coverage.

+++ domain/src/tradex_domain/capabilities.py
@@ -8,6 +8,7 @@ single capability-loud gate used by the SDK services (D-8).
 from __future__ import annotations
 
 from dataclasses import dataclass
+from decimal import Decimal
 
 from tradex_domain.enums import AssetClass
 from tradex_domain.errors import CapabilityNotSupportedError
@@ -36,7 +37,7 @@ class BrokerCapabilities:
     supports_news: bool = False
     supports_fundamentals: bool = False
     max_order_quantity: int | None = None
-    max_order_value: float | None = None
+    max_order_value: Decimal | None = None
     max_batch_size: int = 1
     #: Levels in the deepest market-depth stream (0 = no depth feed).
     #: Dhan depth-20 -> 20; Upstox full_d30 -> 30.
